+++ woocommerce/inc/wc-breadcrumb.php
@@ -1,0 +1,34 @@
+<?php
+
+/**
+ * Breadcrumb
+ *
+ * @package Bootscore 
+ * @version 5.3.3
+ */
+
+
+// Exit if accessed directly
+defined( 'ABSPATH' ) || exit;
+
+
+/**
+ * WooCommerce Breadcrumb
+ */
+if (!function_exists('bs_woocommerce_breadcrumbs')) :
+  add_filter('woocommerce_breadcrumb_defaults', 'bs_woocommerce_breadcrumbs');
+  function bs_woocommerce_breadcrumbs() {
+    return array(
+      'delimiter'   => '',
+      'wrap_before' => "<nav aria-label='breadcrumb' class='wc-breadcrumb breadcrumb-scroller mb-4 mt-2 py-2 px-3 bg-body-tertiary rounded'>
+      <ol class='breadcrumb mb-0'>",
+      'wrap_after'  => '</ol>
+      </nav>',
+      'before'      => '<li class="breadcrumb-item">',
+      'after'       => '</li>',
+      // Remove "Home" and add Fontawesome house icon (_wc_breadcrumb.scss)
+      //'home'        => _x('Home', 'breadcrumb', 'woocommerce'),
+      'home'        => ' ',
+    );
+  }
+endif;
